refactor(doc processor): Move make_pdftotext_process to lib
